api/requirements.txt
-i https://pypi.org/simple
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12 ; python_version >= '3'
click==8.0.4
dataclasses==0.8 ; python_version < '3.7'
flask-cors==3.0.10
flask-socketio==5.1.2
flask==2.0.3
idna==3.3 ; python_version >= '3'
importlib-metadata==4.8.3 ; python_version < '3.8'
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1
numpy==1.19.5
pandas==1.1.5
pymongo==4.1.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1 ; python_version < '3.8'
urllib3==1.26.9
werkzeug==2.0.3
zipp==3.6.0-i https://pypi.org/simple
bidict==0.22.1 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
flask==2.2.3
flask-cors==3.0.10
flask-socketio==5.3.2
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
numpy==1.24.2 ; python_version < '3.10'
pandas==1.5.3
pymongo==4.3.3
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-engineio==4.3.4 ; python_version >= '3.6'
python-socketio==5.7.2 ; python_version >= '3.6'
pytz==2022.7.1
requests==2.28.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.3 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
